pythonFiles/requirements.txt
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cvxpy==1.1.14
cycler==0.10.0
Cython==0.29.24
ecos==2.0.7.post1
Fiona==1.8.21
gambit==16.0.1
geopandas==0.10.2
idna==3.3
kiwisolver==1.3.1
matplotlib==3.4.3
munch==2.5.0
networkx==2.7.1
numpy==1.21.1
osmnx==1.1.2
osqp==0.6.2.post0
pandas==1.4.1
Pillow==8.3.1
pycparser==2.21
pyparsing==2.4.7
pyproj==3.3.0
python-dateutil==2.8.2
python-igraph==0.9.6
pytz==2021.3
qdldl==0.1.5.post0
requests==2.27.1
Rtree==0.9.7
scipy==1.7.0
scs==2.1.4
Shapely==1.8.1.post1
six==1.16.0
texttable==1.6.4
urllib3==1.26.8
